fix: handle edge case in distance to line
Fix incorrect edge case in distance to line calculation: if the points defining the line are coincident, return the distance between points rather than NaN. Update corresponding tests.

Refs: #31
